Whitelabel #1/requirements.txt
python-telegram-bot==20.7
supabase==2.3.0
httpx==0.26.0
websockets==12.0
python-dotenv==1.0.0
aiohttp==3.9.3
pandas==2.2.0
numpy==1.26.4
ta==0.11.0
